requirements.txt
crewai
langchain_groq
langchain
playwright
chromadb
beautifulsoup4
python-dotenv
schedule
python-dotenv
litellm
playwright-stealth
sentence-transformers
fastapi 
uvicorn